Lesson 7 homework level_3

+++ src/main/java/lv/javaguru/java1/student_deniss_boltunovs/lesson_7/homework/level_3/ChequeCalculator.java
@@ -20,7 +20,7 @@ class ChequeCalculator {
             return 0;
         } else {
             double totalAmount = calculateTotalSalesAmount(cheques);
-            return (double) (totalAmount / cheques.length);
+            return (totalAmount / cheques.length);
         }
     }
